+++ CHANGES.rst
@@ -28,6 +28,9 @@ Unreleased
     it's disabled in config. Previously, only disabling worked. :issue:`5916`
 -   ``Flask.select_jinja_autoescape`` uses case-insensitive comparison instead
     of only lower case file extensions. :pr:`6012`
+-   Add ``query`` method route shortcut for the ``QUERY`` HTTP method
+    (:rfc:`10008`). ``MethodView`` dispatches ``QUERY`` requests to a ``query``
+    handler. :issue:`XXXX`
 
 
 Version 3.1.3